Remove hardcoded credentials from Dockerfile
The credentials were removed from the Dockerfile ensuring sensitive data related to login are not exposed. Such a situation could pose a security risk. It is advisable not to hardcode the credentials instead use more secure way like environment variables, which can be fetched during runtime. Prior to this change, the credentials were visible to everyone with access to the file.