Remove default values from interfaces flag
Default values for interfaces flags removed in main.go file. This was done in order to prevent any misunderstanding or misconfiguration issues. The user should manually input the names of the interfaces they wish to monitor for bandwidth scraping. The default values might not always apply to the user's setup which could lead to incorrect performance metrics.